+++ src/InfluxDB2/Client.php
@@ -16,7 +16,7 @@ class Client
     /**
      * Client version updated by: 'make release VERSION=1.5.0'
      */
-    const VERSION = 'dev';
+    const VERSION = '2.7.0';
 
     public $options;
     public $closed = false;